requirements.txt
# Cerelia Production Simulation Tool
# Django Requirements

Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
